update quickstart and add/update/upgrade/reuse tests
